refactor: minor naming cleanups
In this PR I rename `ProxiedNode` as `BenchmarkedNode` because `ProxiedNode` name was just bad (originally discussed [here](https://github.com/AztecProtocol/aztec-packages/pull/19291#discussion_r2662427410)) and I rename `build_nullifier_read_request_hints.ts` to reflect the name of the class it contains.
