+++ yarn-project/pxe/src/private_kernel/hints/index.ts
@@ -1,2 +1,2 @@
-export * from './build_private_kernel_reset_private_inputs.js';
+export * from './private_kernel_reset_private_inputs_builder.js';
 export * from './compute_tx_include_by_timestamp.js';
